collector/pyproject.toml
[project]
name = "default"
version = "0.0.1"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv==1.0.1",
    "pep8-naming==0.14.1",
    "mccabe==0.7.0",
    "pytest==8.3.4",
    "black==25.1.0",
    "aiofiles>=24.1.0",
    "yoyo-migrations>=9.0.0",
    "psycopg[binary]>=3.3.3",
    "psycopg-pool>=3.3.0",
    "cryptography>=46.0.6",
    "pyyaml>=6.0.3",
]

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
# A regex preceded with ^/ will apply only to files and directories
# in the root of the project.
^/foo.py  # exclude a file named foo.py in the root of the project
'''
exclude = '''
/(
    \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | migrations
)/
'''
force-exclude = '''
^/examples/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
